Created generic behaviors file to hold creep behaviors used by multiple roles. Couriers and Harvesters now use these behaviors.
